lambda/social_crawler/requirements.txt
# Core dependencies
requests>=2.31.0
boto3>=1.34.0
psycopg2-binary>=2.9.7
redis>=5.0.0

# Async support
asyncio-throttle>=1.0.2
aiohttp>=3.9.0

# Data processing
python-dateutil>=2.8.2
urllib3>=2.0.0

# Logging and monitoring
structlog>=23.2.0

# Optional: for enhanced text processing
nltk>=3.8.1
textblob>=0.17.1

# Development dependencies (not needed in production)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
